+++ frontend/public/components/nav.jsx
@@ -322,7 +322,6 @@ export const Navigation = ({ isNavOpen, onNavSelect }) => {
           <HrefLink href="/overview" name="Status" activePath="/overview/" required={FLAGS.OPENSHIFT} />
           <HrefLink href="/status" name="Status" activePath="/status/" disallowed={FLAGS.OPENSHIFT} />
           <HrefLink href="/search" name="Search" startsWith={searchStartsWith} />
-          <HrefLink href="/import" name="Import" startsWith={searchStartsWith} />
           <ResourceNSLink resource="events" name="Events" />
         </NavSection>
