+++ frontend/public/style.scss
@@ -90,4 +90,3 @@
 @import "components/storage-class-form";
 @import "components/quota";
 @import "components/cluster-settings/cluster-settings";
-@import "components/import-flow/import-flow"
